Updated formatting in 'Our Experiences' section; added a blank line and removed an extra newline character
The diff shows that there were changes made to the 'Our Experiences' section of the `index.html` file. A blank line was added between the `<h2>` tag and the following content, and an extra newline character was removed after the `<h2>` tag. These changes were made to improve the formatting of the section.